refactor: move Vertex3D to math package
