+++ raytracer/src/main/java/rayrangers/raytracer/math/Vertex3D.java
@@ -1,6 +1,4 @@
-package rayrangers.raytracer.world;
-
-import rayrangers.raytracer.math.Vector3D;
+package rayrangers.raytracer.math;
 
 /**
  * Represents a vertex in 3D space.
