+++ raytracer/src/main/java/rayrangers/raytracer/world/Camera.java
@@ -1,0 +1,6 @@
+package rayrangers.raytracer.world;
+
+public class Camera {
+
+        
+}
